tailq_sort: Generic sort routine for BSD-style TAILQ
This is an implementation of (stable) merge sort for the doubly-linked
lists provided by <sys/queue.h> (or sometimes <bsd/sys/queue.h>),
called tail queues. The implementation follows
http://www.chiark.greenend.org.uk/~sgtatham/algorithms/listsort.html.

It relies on the gcc extension allowing a variable-length array inside
a structure: Inside the implementation of tailq_sort(), I define a
mock-up structure with a TAILQ_ENTRY at the same location as in the
user's struct (using the offset which is passed to the function), and
operate on that.

I chose this evil over my first implementation, which was a
header-only solution: Defining the two macros
DECLARE_TAILQ_SORT_FUNCTION(name, headtype) and
DEFINE_TAILQ_SORT_FUNCTION(name, headtype, type, member), where the
expansion of the latter would be the entire definition of the sort
function. That approach had the advantage of being written in standard
C and providing more type-safety, but requires the user to define a
sort function for each type that needs to be sorted. Also, I really
don't like 70+ line macros.

There is also an interactive test program. Mainly for use with that
test program, I also included a tailq_shuffle() routine; I doubt it
has (m)any practical uses.