jenkins_hash: Further endian simplifications
Since I decided to only define and expose hash_xxx on xxx-endian
machines, there's really no point in keeping the names hashlittle and
hashbig. So simply always provide two functions, hash and hash2, where
only the implementation depends on the endianness of the machine. This
removes the ugly endian-detection from the header.