Simple graph utilities
These are simple utilities I wrote for computing the set of connected
components of a graph and for finding all maximal cliques.

Because I only needed this for a graph read from stdin, the graph
implementation is "append-only", meaning that nodes and edges can be
added but not removed. This in particular simplifies finding the
connected components, since one can simply keep track of those as
edges are added.

graphcomponents is a rather trivial front-end; all the work is already
done while the graph is being built in memory.

maximal_cliques has worked reasonably well for the tasks I originally
wrote it for, but there may still be room for improvement. But before
trying that, I need to find some appropriate data to test it on. Also,
obviously, I'd really like to add some correctness tests.