split_col: Eliminate duplicate code in open_* subs
The 'try to open and close some fds if it fails with EMFILE or ENFILE'
logic was needlessly duplicated in open_read, open_write and
open_append. Create a single sub open_common for that. The old subs
are then mostly trivial wrappers (we still need to write the header on
first open).