random_subset: Pick random lines from file
It is occasionally useful to have a simple tool for picking a random
subset of lines from one or more files. While one can use a
combination of sort -R and head, this is rather inefficient, and
moreover one might want to preserve the order of the lines.

random_subset can operate in two modes: Either an exact number of
lines (or as many as the input provides) if given the -n parameter, or
a certain fraction of the input if given the -f option.

If no options are given, defaults to -n 1, that is, selects a single
line randomly from the input.

If both -f and -n are given, -f takes precedence, but the count is
used as an upper bound for the number of lines printed. So to get
about 2% of all lines but at most 1000, one can use -f 0.02 -n 1000.

The -f implementation is obvious: For each line read, pick a number
uniformly from [0, 1] (as well as perl can provide such a number), and
print the line if the number is less than the given fraction.

For -n K, we would like that for an input of N lines, each of the
binom(N, K) possible subsets are (roughly) equally likely to be
selected. However, we don't know N in advance, and we certainly don't
want to soak up the entire input to find it before selecting the
subset. So we use a little trick (I'm sure I'm not the first to come
up with this):

Assume we've read N >= K lines and we've retained a subset of K lines,
each of the binom(N, K) subsets being equally likely. Now it turns out
that there's an N+1st line of input. To produce a new K-element
subset, we need to decide whether to keep this new line, and if so,
what old line to discard. Now, among all K-element subsets of N+1
elements, the N+1st is a member of precisely K/(N+1) of them. So we
can roll an (N+1)-sided die, and if it shows between 1 and K pips, we
keep the new element, otherwise we discard it. If we keep it, it is
not hard to see that we should roll a K-sided die to decide which old
element to evict. This way, we get a new subset chosen uniformly at
random among all the binom(N+1, K) K-subsets of the input seen so
far. At least when our X-sided dice actually produce elements
uniformly at random. For now, I haven't bothered dealing with the
modulo bias and all the other stuff making random numbers hard.

Instead of splicing the old element out and push the new at the end,
we just overwrite the old element. In order to print the lines in the
original order, we have to keep track of the original line number, and
then we might as well provide an option to prepend that number to the
output. This could also be useful in the -f case, e.g. if
random_subset is used to filter an "infinite" stream, one can roughly
follow the progress of whatever produces that stream by looking at the
line numbers.