split_col: refactor template computation and use
Splitting the template in a prefix and suffix is insufficient for
allowing multiple replacement strings. So update compute_outfn to
always do substitution on the "current template"; the current template
is either the template given on the command line, or the one
implicitly produced from the current input file name.