split_col: enable a few other replacement strings
Inspired by GNU Parallel, we can easily allow {.} and {..} to stand
for the current input file sans extension and the extension,
respectively. In fact, this also allows us to simplify the code
somewhat, since these are precisely what is needed to describe our
default template. Hence we can now just always call
compute_current_template unconditionally when handling a new input
file.

This also fixes a bug compute_current_template(): we should only look
for a . in the last path component.

We can later add support for {/}, {/.} etc. if we feel like it.