split_col: factor out replacement logic
To allow replacement strings {1}, {2} etc., it is useful to lower the
extraction of the relevant column from do_input_file to compute_outfn,
and in fact from there to a separate helper. We pass an array ref with
all the fields all the way through, but to produce useful diagnostics
in case of too few columns we need to stash the current input file in
a global variable.

Since we do that anyway, we can avoid passing it explicitly to
compute_current_template.