split_col: stash fields in global variable
Parameter passing in perl is a little expensive, unfortunately. So
instead of passing a reference to @fields all the way through to
get_replacement, just store an anonymous array ref in the global
$current_fields and use that. A simple test on a file with ~1 million
lines now runs in about 10.3 seconds, whereas previously it took 11.4
seconds. That's 10%, and for the amount of data I shuffle around,
that's certainly worth it.