split_col: enable {n} replacement strings
We can now easily allow splitting on multiple columns, while at the
same time simplifying the user interface. We simply allow replacement
strings {n} for positive integers n. This mostly makes the --column
option redundant, but we keep it for backwards compatibility. {} will
continue to correspond to {c}, where c is the value of the --column
option.